fix & optz

+++ src/utils/color.cc
@@ -27,28 +27,28 @@
 
 namespace {
 enum ConsoleColor : uint16_t {
-    fBlue = FOREGROUND_BLUE,
-    fGreen = FOREGROUND_GREEN,
-    fRed = FOREGROUND_RED,
+    fBlue      = FOREGROUND_BLUE,
+    fGreen     = FOREGROUND_GREEN,
+    fRed       = FOREGROUND_RED,
     fIntensity = FOREGROUND_INTENSITY,
 
-    bBlue = BACKGROUND_BLUE,            // 0x10
-    bGreen = BACKGROUND_GREEN,          // 0x20
-    bRed = BACKGROUND_RED,              // 0x40
+    bBlue      = BACKGROUND_BLUE,       // 0x10
+    bGreen     = BACKGROUND_GREEN,      // 0x20
+    bRed       = BACKGROUND_RED,        // 0x40
     bIntensity = BACKGROUND_INTENSITY,  // 0x80
 
     // Foreground color
-    fYellow = fRed | fGreen,
-    fCyan = fGreen | fBlue,
+    fYellow  = fRed | fGreen,
+    fCyan    = fGreen | fBlue,
     fMagenta = fBlue | fRed,
-    fWhite = fRed | fBlue | fGreen,
+    fWhite   = fRed | fBlue | fGreen,
     fDefault = fWhite,
 
     // Background color
-    bYellow = bRed | bGreen,
-    bCyan = bGreen | bBlue,
+    bYellow  = bRed | bGreen,
+    bCyan    = bGreen | bBlue,
     bMagenta = bBlue | bRed,
-    bWhite = bRed | bBlue | bGreen,
+    bWhite   = bRed | bBlue | bGreen,
     bDefault = 0,  // 0 black
 
     defColor = fDefault | bDefault,
@@ -95,7 +95,8 @@ void SetConsoleColor(FILE *f, Color foreground, Color background) {
     } else {
         handle = GetStdHandle(STD_ERROR_HANDLE);
     }
-    WORD attr = ForegroundColorTable[foreground] + BackgroundColorTable[background];
+    WORD attr = ForegroundColorTable[static_cast<int>(foreground)] +
+                BackgroundColorTable[static_cast<int>(background)];
     if (win32_reset_color == -1) {
         win32_reset_color = GetConsoleColor(handle);
     }
@@ -125,24 +126,24 @@ void FprintColorTextLine(FILE *f, Color foreground, Color background, const char
 #else
 namespace {
 
-const char *fBlack = "\033[30m";
-const char *fRed = "\033[31m";
-const char *fGreen = "\033[32m";
-const char *fYellow = "\033[33m";
-const char *fBlue = "\033[34m";
+const char *fBlack   = "\033[30m";
+const char *fRed     = "\033[31m";
+const char *fGreen   = "\033[32m";
+const char *fYellow  = "\033[33m";
+const char *fBlue    = "\033[34m";
 const char *fMagenta = "\033[35m";
-const char *fCyan = "\033[36m";
-const char *fWhite = "\033[37m";
+const char *fCyan    = "\033[36m";
+const char *fWhite   = "\033[37m";
 const char *fDefault = "\033[39m";  // default foreground color
 
-const char *bBlack = "\033[40m";
-const char *bRed = "\033[41m";
-const char *bGreen = "\033[42m";
-const char *bYellow = "\033[43m";
-const char *bBlue = "\033[44m";
+const char *bBlack   = "\033[40m";
+const char *bRed     = "\033[41m";
+const char *bGreen   = "\033[42m";
+const char *bYellow  = "\033[43m";
+const char *bBlue    = "\033[44m";
 const char *bMagenta = "\033[45m";
-const char *bCyan = "\033[46m";
-const char *bWhite = "\033[47m";
+const char *bCyan    = "\033[46m";
+const char *bWhite   = "\033[47m";
 const char *bDefault = "\033[49m";  // default background color
 
 const char *colorReset = "\033[0m";
@@ -158,7 +159,8 @@ const char *BackgroundColorTable[] = {
 
 namespace raptor {
 void SetConsoleColor(FILE *f, Color foreground, Color background) {
-    fprintf(f, "%s%s", ForegroundColorTable[foreground], BackgroundColorTable[background]);
+    fprintf(f, "%s%s", ForegroundColorTable[static_cast<int>(foreground)],
+            BackgroundColorTable[static_cast<int>(background)]);
 }
 
 void ResetConsoleColor(FILE *f) {
@@ -167,8 +169,8 @@ void ResetConsoleColor(FILE *f) {
 
 void FprintColorTextLine(FILE *f, Color foreground, Color background, const char *text) {
     SetConsoleColor(f, foreground, background);
-    fprintf(f, "%s%s%s%s\n", ForegroundColorTable[foreground], BackgroundColorTable[background],
-            text, colorReset);
+    fprintf(f, "%s%s%s%s\n", ForegroundColorTable[static_cast<int>(foreground)],
+            BackgroundColorTable[static_cast<int>(background)], text, colorReset);
     ResetConsoleColor(f);
 }
 }  // namespace raptor
